backend/requirements.txt
# Core Flask and Web Framework
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-SocketIO>=5.3.0
python-socketio>=5.8.0
python-engineio>=4.7.0

# AI and Machine Learning
openai>=1.3.0
google-generativeai>=0.3.0

# Speech Processing
SpeechRecognition>=3.10.0
pyttsx3>=2.90
gTTS>=2.3.0

# HTTP and API Requests
requests>=2.31.0
aiohttp>=3.8.0
httpx>=0.25.0

# Environment and Configuration
python-dotenv>=1.0.0

# Real-time Communication
websockets>=11.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Web Scraping and Search
beautifulsoup4>=4.12.0
google-api-python-client>=2.108.0

# Weather and Location
pyowm>=3.3.0
geopy>=2.4.0

# System and Utilities
psutil>=5.9.0

# Additional Utilities
Pillow>=10.1.0
python-dateutil>=2.8.0
pytz>=2023.3 